moved h1 styled to global styles, created Input and CTAButton components
