+++ src/components/CTAButton/CTAButton.js
@@ -1,0 +1,1 @@
+export { CTAButton as default } from './CTAButton.styles';
